% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelectoralbar.R
\name{allelectoralbar}
\alias{allelectoralbar}
\title{Generates a bar graph of the winning percentage of electoral votes ordered by year.}
\usage{
allelectoralbar()
}
\examples{
#' @title Generates a bar graph of the winning percentage of electoral votes ordered by year.
#' @aliases allelectoralbar
#' @example
#' R/allelectoralbar.R

allelectoralbar <- function(){
  require(grDevices)
  votes <- read.csv("VotesData.csv", stringsAsFactors = FALSE)
  percents <- data.frame(votes[2], votes[4])
  plot(percents, type = "h", col = "blue", 
       main = "Electoral Vote Won by Year", 
       xlab = "Year", ylab = "Percent of Votes")
}
}

