% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpopularbar.R
\name{allpopularbar}
\alias{allpopularbar}
\title{allpopularbar}
\usage{
allpopularbar()
}
\description{
Generates a bar graph of the winning percentage of popular votes ordered by year.
}
\details{
A bar graph showing the percentage of popular votes won by each president for each election year from 1828-2012 is generated.
}
\examples{
#' @title allpopularbar
#' @description Generates a bar graph of the winning percentage of popular votes ordered by year.
#' @details A bar graph showing the percentage of popular votes won by each president for each election year from 1828-2012 is generated.
#' @example 
#' R/allpopularbar.R

allpopularbar <- function(){
  require(grDevices)
  votes <- read.csv("VotesData.csv", stringsAsFactors = FALSE)
  percents <- data.frame(votes[2:3])
  plot(percents, type = "h", col = "blue", 
      main = "Popular Vote Won by Year", 
      xlab = "Year", ylab = "Percent of Votes")
}
}

