% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralbar.R
\name{electoralbar}
\alias{electoralbar}
\title{Generates a bar graph of the winning percentage of electoral votes for the given election year.}
\usage{
electoralbar(year)
}
\arguments{
\item{year}{An integer}
}
\examples{
/R/electoralbar.R
}

