% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularbar.R
\name{popularbar}
\alias{popularbar}
\title{Generates a bar graph of the winning percentage of popular votes for the given election year.}
\usage{
popularbar(year)
}
\arguments{
\item{year}{An integer}
}
\examples{
R/popularbar.R
}

