% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnoutlg.R
\docType{data}
\name{turnout}
\alias{turnout}
\title{turnoutlg}
\format{An object of class \code{data.frame} with 47 rows and 2 columns.}
\usage{
turnout
}
\description{
Generates a line graph of the voter turnout data.
}
\details{
A line graph showing voter turnout from 1828-2012 is generated.
}
\keyword{datasets}

